MediMeowBackend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1
aiofiles==23.2.1
pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2
